requirements.txt
requests~=2.32.3
pillow~=10.4.0
python-dotenv~=1.0.1
openai~=1.43.1
SQLAlchemy~=2.0.34
ttkbootstrap~=1.10.1
exceptiongroup~=1.2.2
colorlog~=6.9.0